test/server: demonstrate inaccurate QoS with multiple queues per server
Add a couple of tests using immediate and delayed tag calculation to
demonstrate inaccurate QoS provided to a client when multiple queues are
configured per server. The requests from the client are distributed across
multiple queues. Since the queues are independent, the tag calculations
are also independent and this results in inaccurate QoS provided to the
client as the tests show. The following scenario outlines problem with
multiple queues per server.

Tag Calculation with Single Queue:
----------------------------------

 Enqueue--->|T5|T4|T3|T2|T1|T0|--->Dequeue
         Tags for items in Single Queue

  Where, Tn = Tag value at time = n

For simplicity, consider x to be any of the
client info parameters i.e. reservation, weight or limit

1. Tag calculation for T1 in the queue would look like:

    T1 = max((T0 + (1/x)), arr_time)
    where, T0 - Tag at time 0
           x  - client info [res|wgt|lim], and,
           arr_time - Arrival time of the request in the tag

2. Similarly, Tag calculation for T5:

    T5 = max((T4 + (1/x)), arr_time)

As seen above, there is no issue when a single mClock
queue is in operation. For all the requests, the tags are
calculated based on their arrival times and in sequence.

Tag Calculation with Multiple Queues per Server:
-----------------------------------------------
Consider two queues configured on the same server with
items added in the following sequence:

   Enqueue--->|T5|T2|T0|--->Dequeue
                   Queue0

   Enqueue--->|T4|T3|T1|--->Dequeue
                   Queue1

Consider the request arrival times in each queue according
to the number associated with each tag, for example:

 - Req0 arrives on queue0 at time t0
 - Req1 arrives on queue1 at time t1
 - Req2 arrives on queue0 at time t2
and so on with the final Req5 arriving on queue0 at time t5

Tag calcs on Queue0                 | Tag calcs on Queue1
------------------------------------|---------------------------------
T0 = max((NO_TAG + (1/x)), arr_time)| T1 = max((NO_TAG + (1/x)), arr_time)
T2 = max((T0 + (1/x)), arr_time)    | T3 = max((T1 + (1/x)), arr_time)
T5 = max((T2 + (1/x)), arr_time)    | T4 = max((T3 + (1/x)), arr_time)

For simplicity, assume reservation and limit is set to 3 IOPS, AtLimit is
set to 'Wait' and all requests arrived well within a few milliseconds on
both queues.

It's pretty clear that all the 3 requests from both the queues will be
scheduled resulting in 6 IOPS which is not correct.

Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>