server: Setup the necessary infrastructure to enable dynamic tag info
The groundwork to set up dynamic tag info is similar to dynamic client
info with a key difference. The client info relies on only one function
to enable clients to update the client info parameters. But dynamic tag
info via is_dynamic_tag_info_f relies on two functions outlined below
to enable both the client and server to update and get the necessary
information to calculate the next accurate tag. The client of dmClock is
expected to maintain a map of clientId to ReqTagInfo similar to
ClientInfo.

The two dynamic tag info functions are:

 1) reqtag_updt_f: The server uses this function to update the latest
    calculated tag to the dmClock client. This is called whenever a new
    tag is calculated either as part of the initial_tag() or
    update_next_tag() depending on the type tag calculation employed.

    The client on the other hand updates the 'last_tick_interval' in the
    map just before adding a new request to the queue.

 2) reqtag_info_f: The server uses this function to get the latest tag
    for a clientId by reading the 'last_tag' and the latest tick interval
    (for DelayedTagCalc) via 'last_tick_interval'. This information is
    used to calculate the next accurate tag.

To enable the above functionality, additional PriorityQueueBase and
PullPriorityQueue constructors are defined for dmClock clients to
appropriately set up the dynamic functions and the associated data
structures. A bool U2 (false by default) is used to tell the server to
enable dynamic tag info functionality. This is propagated accordingly to
both the PullPriorityQueue and PushPriorityQueue constructors.

However, it is important to note that the dynamic tag info functionality
is currently only implemented for PullPriorityQueue.

Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>